fix: simplify, remove shared file select fragment
